Drar inn nye Organisasjonsnummer-typen
- Og starter da også bruken av optional for å slippe Null-typer for `EbmsAktoer` og `Organisasjonsnummer`.
- Oppgraderer animal-sniffer til å sjekke for java 8-kompatibilitet
