+++ api-commons/src/main/java/no/digipost/api/interceptors/TransactionLog.java
@@ -143,7 +143,7 @@ public class TransactionLog {
 
 	private String getOrgNr(final EbmsContext context) {
 		StringBuilder builder = new StringBuilder();
-		if (context.remoteParty != null && context.remoteParty != Organisasjonsnummer.NULL) {
+		if (context.remoteParty != null && context.remoteParty.isPresent()) {
 			builder.append(context.remoteParty.toString());
 		} else {
 			builder.append("-");
