+++ api-commons/src/main/java/no/digipost/api/representations/Organisasjonsnummer.java
@@ -16,75 +16,24 @@
 package no.digipost.api.representations;
 
 import no.digipost.api.PMode;
-import org.apache.commons.lang3.ArrayUtils;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+public interface Organisasjonsnummer extends MedOrganisasjonsnummer{
 
-public class Organisasjonsnummer {
+	String ISO6523_ACTORID = PMode.PARTY_ID_TYPE;
 
-	public static final Pattern ISO6523_PATTERN = Pattern.compile("^([0-9]{4}:)?([0-9]{9})$");
-	public static final Organisasjonsnummer NULL = new Organisasjonsnummer("");
-	public static final String ISO6523_ACTORID = PMode.PARTY_ID_TYPE;
-
-	private static final String COUNTRY_CODE_ORGANIZATION_NUMBER_NORWAY = "9908";
-
-	private final String organisasjonsnummer;
-
-	private Organisasjonsnummer(final String organisasjonsnummer) {
-		this.organisasjonsnummer = organisasjonsnummer;
+	static Organisasjonsnummer of(String organisasjonsnummer){
+		return new EtOrganisasjonsnummer(organisasjonsnummer);
 	}
 
-	public String medLandkode() {
-		return String.format("%s:%s", COUNTRY_CODE_ORGANIZATION_NUMBER_NORWAY, organisasjonsnummer);
-	}
+	DatabehandlerOrganisasjonsnummer forfremTilDatabehandler();
 
-	public String utenLandkode() {
-		return organisasjonsnummer;
-	}
-
-	public static boolean isOrganisasjonsnummer(final String organisasjosnummer) {
-		return ISO6523_PATTERN.matcher(organisasjosnummer).matches();
-	}
-
-	public static Organisasjonsnummer of(final String organisasjonsnummer) {
-		Matcher matcher = ISO6523_PATTERN.matcher(organisasjonsnummer);
-		if (!matcher.matches()) {
-			throw new IllegalArgumentException(
-					String.format("Ugyldig organisasjonsnummer. Forventet format er ISO 6523, men fikk følgende nummer: '%s'. " +
-							"Organisasjonsnummeret skal være 9 siffer og kan prefikses med landkode 9908. " +
-							"Eksempler på dette er '9908:984661185' og '984661185'.",
-							organisasjonsnummer)
-			);
-		}
-		return new Organisasjonsnummer(matcher.group(2));
-	}
+	AvsenderOrganisasjonsnummer forfremTilAvsender();
 
 	@Override
-	public String toString() {
-		return utenLandkode();
-	}
+	String getOrganisasjonsnummerMedLandkode();
 
-    @Override
-	public boolean equals(final Object obj) {
-		if (obj == null) {
-			return false;
-		}
-
-		if (obj instanceof Organisasjonsnummer) {
-			return organisasjonsnummer.equals(((Organisasjonsnummer)obj).organisasjonsnummer);
-		}
-		return false;
-	}
-
-    @Override
-	public int hashCode() {
-		return organisasjonsnummer.hashCode();
-	}
-
-	public boolean oneOf(Organisasjonsnummer ... candidates) {
-		return ArrayUtils.contains(candidates, this);
-	}
+	@Override
+	String getOrganisasjonsnummerUtenLandkode();
 
 
 }
