--- api-commons/src/main/java/no/digipost/api/representations/OrgnrNEW.java
@@ -1,39 +1,0 @@
-/**
- * Copyright (C) Posten Norge AS
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package no.digipost.api.representations;
-
-import no.digipost.api.PMode;
-
-public interface OrgnrNEW extends MedOrganisasjonsnummer{
-
-	String ISO6523_ACTORID = PMode.PARTY_ID_TYPE;
-
-	static OrgnrNEW of(String organisasjonsnummer){
-		return new EtOrganisasjonsnummer(organisasjonsnummer);
-	}
-
-	DatabehandlerOrganisasjonsnummer forfremTilDatabehandler();
-
-	AvsenderOrganisasjonsnummer forfremTilAvsender();
-
-	@Override
-	String getOrganisasjonsnummerMedLandkode();
-
-	@Override
-	String getOrganisasjonsnummerUtenLandkode();
-
-
-}
