+++ pom.xml
@@ -350,11 +350,15 @@
             <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.3</version>
+                <configuration>
+                    <source>1.8</source>
+                    <target>1.8</target>
+                </configuration>
             </plugin>
             <plugin>
             	<groupId>org.codehaus.mojo</groupId>
             	<artifactId>animal-sniffer-maven-plugin</artifactId>
-            	<version>1.14</version>
+            	<version>1.7</version>
             	<executions>
             		<execution>
             			<id>ensure-java6-compatible</id>
@@ -365,8 +369,8 @@
             			<configuration>
             				<signature>
             					<groupId>org.codehaus.mojo.signature</groupId>
-                				<artifactId>java16</artifactId>
-                				<version>1.1</version>
+                				<artifactId>java18</artifactId>
+                				<version>1.0</version>
             				</signature>
             			</configuration>
             		</execution>
